Add functionality for storing data
This PR includes functionality to store and export a chunk of data (155 bytes), which is separated from the rest of the applet. This gives us a bit of future proofing, allowing us to support Shamir splitting and multisig operations in the future. Both APDUs are PIN protected and use secure channels.

The data is 155 bytes because that allows us to store up to 5 compressed public keys (33 bytes each) for a Bitcoin multisig. I believe this is sufficient splitting for most purposes and allows us to add more data-intensive features to v2+ cards (e.g. Phonon) while maintaining backward compatibility.